Rename usb_request.h to usb_types.h
In the future all shared USB data types can live in this file.
